+++ pcs/lib/cib/description.py
@@ -1,0 +1,55 @@
+from lxml.etree import _Element
+
+from pcs.common import reports
+from pcs.lib.cib.const import (
+    TAG_ACL_PERMISSION,
+    TAG_ACL_ROLE,
+    TAG_ALERT,
+    TAG_LIST_RESOURCE,
+    TAG_NODE,
+    TAG_RECIPIENT,
+)
+from pcs.lib.xml_tools import update_attribute_remove_empty
+
+TAG_LIST_SUPPORTS_DESCRIPTION = frozenset.union(
+    TAG_LIST_RESOURCE,
+    {
+        TAG_ACL_PERMISSION,
+        TAG_ACL_ROLE,
+        TAG_ALERT,
+        TAG_NODE,
+        TAG_RECIPIENT,
+    },
+)
+
+DESCRIPTION_ATTRIBUTE = "description"
+
+
+def validate_description_support(element: _Element) -> reports.ReportItemList:
+    """
+    Validate that element supports the description attribute
+
+    element -- specified element
+    """
+    if element.tag not in TAG_LIST_SUPPORTS_DESCRIPTION:
+        return [
+            reports.ReportItem.error(
+                reports.messages.IdBelongsToUnexpectedType(
+                    str(element.attrib["id"]),
+                    sorted(TAG_LIST_SUPPORTS_DESCRIPTION),
+                    current_type=element.tag,
+                )
+            )
+        ]
+    return []
+
+
+def set_description(element: _Element, description: str) -> None:
+    """
+    Set the description of element to specified value or remove the description
+    attribute if the new value is empty
+
+    element -- specified element
+    description -- new description for the element
+    """
+    update_attribute_remove_empty(element, DESCRIPTION_ATTRIBUTE, description)